core/skills/deepcamera/skills/transformation/depth-estimation/requirements.txt
# Depth Estimation — Privacy Transform Skill
# CoreML-first on macOS (Neural Engine), PyTorch fallback on other platforms.
#
# macOS: coremltools loads .mlpackage models — fast, leaves GPU free.
# Other: PyTorch + depth-anything-v2 pip package + HF weights.
# Common: opencv, numpy, pillow, huggingface_hub for model download.

# ── CoreML (macOS only) ──────────────────────────────────────────────
coremltools>=8.0; sys_platform == "darwin"

# ── PyTorch fallback (non-macOS, or if CoreML unavailable) ───────────
# NOTE: torch and torchvision MUST be version-paired.
torch~=2.7.0
torchvision~=0.22.0
depth-anything-v2>=0.1.0

# ── Common dependencies ─────────────────────────────────────────────
huggingface_hub>=0.20.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0

# ── TensorRT (optional, Windows/Linux NVIDIA) ────────────────────────
# If available, transform.py auto-selects TRT FP16 for ~7x speedup.
# Falls back to PyTorch CUDA if not installed.
tensorrt>=10.0; sys_platform != "darwin"
